10.09

+++ src/main/java/com/yupi/springbootinit/common/PageRequest.java
@@ -1,13 +1,13 @@
 package com.yupi.springbootinit.common;
 
+
 import com.yupi.springbootinit.constant.CommonConstant;
 import lombok.Data;
 
 /**
  * 分页请求
  *
- *  
- *   
+ * 
  */
 @Data
 public class PageRequest {
@@ -15,12 +15,12 @@ public class PageRequest {
     /**
      * 当前页号
      */
-    private int current = 1;
+    private long current = 1;
 
     /**
      * 页面大小
      */
-    private int pageSize = 10;
+    private long pageSize = 10;
 
     /**
      * 排序字段
